@@ -12,7 +12,8 @@ import math
 from decimal import Decimal
 from typing import Any, Union, SupportsFloat
 
-from ..helpers import collapse_white_spaces
+from ..helpers import BOOLEAN_VALUES, NUMERIC_INF_OR_NAN, INVALID_NUMERIC, \
+    collapse_white_spaces
 from .atomic_types import AtomicTypeMeta
 from .untyped import UntypedAtomic
 from .numeric import Float10, Integer
@@ -41,7 +42,7 @@ class BooleanProxy(metaclass=AtomicTypeMeta):
         elif not isinstance(value, str):
             raise TypeError('invalid type {!r} for xs:{}'.format(type(value), cls.name))
 
-        if value.strip() not in {'true', 'false', '1', '0'}:
+        if value.strip() not in BOOLEAN_VALUES:
             raise ValueError('invalid value {!r} for xs:{}'.format(value, cls.name))
         return 't' in value or '1' in value
 
@@ -106,10 +107,9 @@ class DoubleProxy10(metaclass=AtomicTypeMeta):
     def __new__(cls, value: Union[SupportsFloat, str]) -> float:  # type: ignore[misc]
         if isinstance(value, str):
             value = collapse_white_spaces(value)
-            if value in {'INF', '-INF', 'NaN'} or cls.xsd_version != '1.0' and value == '+INF':
+            if value in NUMERIC_INF_OR_NAN or cls.xsd_version != '1.0' and value == '+INF':
                 pass
-            elif value.lower() in {'inf', '+inf', '-inf', 'nan',
-                                   'infinity', '+infinity', '-infinity'}:
+            elif value.lower() in INVALID_NUMERIC:
                 raise ValueError('invalid value {!r} for xs:{}'.format(value, cls.name))
         return float(value)
 
