@@ -7,7 +7,7 @@
 #
 # @author Davide Brunato <brunato@sissa.it>
 #
-from typing import cast, Any, Optional
+from typing import Any, Optional
 
 from ..helpers import QNAME_PATTERN
 from .atomic_types import AtomicTypeMeta
@@ -26,7 +26,7 @@ class AbstractQName(metaclass=AtomicTypeMeta):
     def __new__(cls, *args: Any, **kwargs: Any) -> 'AbstractQName':
         if cls.__name__ == 'Notation':
             raise TypeError("can't instantiate xs:NOTATION objects")
-        return cast(AbstractQName, super().__new__(cls))
+        return super().__new__(cls)
 
     def __init__(self, uri: Optional[str], qname: str) -> None:
         if uri is None:
@@ -57,6 +57,10 @@ class AbstractQName(metaclass=AtomicTypeMeta):
     @property
     def expanded_name(self) -> str:
         return '{%s}%s' % (self.uri, self.local_name) if self.uri else self.local_name
+
+    @property
+    def braced_uri_name(self) -> str:
+        return 'Q{%s}%s' % (self.uri, self.local_name) if self.uri else self.local_name
 
     def __repr__(self) -> str:
         return '%s(uri=%r, qname=%r)' % (self.__class__.__name__, self.uri, self.qname)
