@@ -17,7 +17,7 @@ from typing import overload, Any
 if sys.version_info < (3, 8):
     # for Python < 3.8 fallback to typing.Any
     ElementProtocol = Any
-    LxmlElementProtocol: Any
+    LxmlElementProtocol = Any
     DocumentProtocol = Any
     XsdComponentProtocol = Any
     XsdTypeProtocol = Any
@@ -50,7 +50,11 @@ else:
 
     @runtime_checkable
     class LxmlElementProtocol(ElementProtocol, Protocol):
-        def getparent(self) -> Optional[Any]: ...
+        def getnext(self) -> Optional['LxmlElementProtocol']: ...
+        def getparent(self) -> Optional['LxmlElementProtocol']: ...
+        def getprevious(self) -> Optional['LxmlElementProtocol']: ...
+        def itersiblings(self, tag: Optional[str] = ..., preceding: bool = False,
+                         *tags: str) -> Iterable['LxmlElementProtocol']: ...
         nsmap: Dict[Optional[str], str]
 
     class DocumentProtocol(Iterable[ElementProtocol], Hashable, Protocol):
@@ -68,48 +72,69 @@ else:
     class XsdTypeProtocol(XsdComponentProtocol, Protocol):
         def is_simple(self) -> bool:
             """Returns `True` if it's a simpleType instance, `False` if it's a complexType."""
+            ...
+
         def is_empty(self) -> bool:
             """
             Returns `True` if it's a simpleType instance or a complexType with empty content,
             `False` otherwise.
             """
+            ...
+
         def has_simple_content(self) -> bool:
             """
             Returns `True` if it's a simpleType instance or a complexType with simple content,
             `False` otherwise.
             """
+            ...
+
         def has_mixed_content(self) -> bool:
             """
             Returns `True` if it's a complexType with mixed content, `False` otherwise.
             """
+            ...
+
         def is_element_only(self) -> bool:
             """
             Returns `True` if it's a complexType with element-only content, `False` otherwise.
             """
+            ...
+
         def is_key(self) -> bool:
             """Returns `True` if it's a simpleType derived from xs:ID, `False` otherwise."""
+            ...
+
         def is_qname(self) -> bool:
             """Returns `True` if it's a simpleType derived from xs:QName, `False` otherwise."""
+            ...
+
         def is_notation(self) -> bool:
             """Returns `True` if it's a simpleType derived from xs:NOTATION, `False` otherwise."""
+            ...
+
         def is_valid(self, obj: Any, *args: Any, **kwargs: Any) -> bool:
             """
             Validates an XML object node using the XSD type. The argument *obj* is an element
             for complex type nodes or a text value for simple type nodes. Returns `True` if
             the argument is valid, `False` otherwise.
             """
+            ...
+
         def validate(self, obj: Any, *args: Any, **kwargs: Any) -> None:
             """
             Validates an XML object node using the XSD type. The argument *obj* is an element
             for complex type nodes or a text value for simple type nodes. Raises a `ValueError`
             compatible exception (a `ValueError` or a subclass of it) if the argument is not valid.
             """
+            ...
+
         def decode(self, obj: Any, *args: Any, **kwargs: Any) -> Any:
             """
             Decodes an XML object node using the XSD type. The argument *obj* is an element
             for complex type nodes or a text value for simple type nodes. Raises a `ValueError`
             or a `TypeError` compatible exception if the argument it's not valid.
             """
+            ...
 
     class XsdElementProtocol(XsdComponentProtocol, ElementProtocol, Protocol):
         type: XsdTypeProtocol
