@@ -12,6 +12,6 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from .xpath2_parser import XPath2Parser
 else:
-    from .xpath2_constructors import XPath2Parser
+    from ._xpath2_constructors import XPath2Parser
 
 __all__ = ['XPath2Parser']
