@@ -1,5 +1,5 @@
 #
-# Copyright (c), 2018-2020, SISSA (International School for Advanced Studies).
+# Copyright (c), 2018-2021, SISSA (International School for Advanced Studies).
 # All rights reserved.
 # This file is distributed under the terms of the MIT License.
 # See the file 'LICENSE' in the root directory of the present
@@ -7,6 +7,11 @@
 #
 # @author Davide Brunato <brunato@sissa.it>
 #
-from .xpath31_parser import XPath31Parser
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from .xpath31_parser import XPath31Parser
+else:
+    from ._xpath31_functions import XPath31Parser
 
 __all__ = ['XPath31Parser']
