@@ -150,7 +150,12 @@ class XPath2ConstructorsTest(xpath_test_class.XPathTestCase):
         self.check_value('xs:QName(())', [])
         self.check_value('xs:QName("xs:element")', qname)
         self.check_value('xs:QName(xs:QName("xs:element"))', qname)
-        self.check_value('xs:QName(xs:untypedAtomic("xs:element"))', qname)
+
+        if self.parser.version == '2.0':
+            self.wrong_type('xs:QName(xs:untypedAtomic("xs:element"))', 'XPTY0004')
+        else:
+            self.check_value('xs:QName(xs:untypedAtomic("xs:element"))', qname)
+
         self.wrong_type('xs:QName(5)', 'XPTY0004', "the argument has an invalid type")
         self.wrong_value('xs:QName("1")', 'FORG0001', "invalid value")
 
