@@ -440,8 +440,8 @@ class XPath2ParserTest(test_xpath1_parser.XPath1ParserTest):
         self.check_value("5 ge 9", False)
         self.check_value("5 gt 3", True)
         self.check_value("5 lt 20.0", True)
-        self.check_value("false() eq 1", False)
-        self.check_value("0 eq false()", True)
+        self.wrong_type("false() eq 1", 'XPTY0004')
+        self.wrong_type("0 eq false()", 'XPTY0004')
         self.check_value("2 * 2 eq 4", True)
         self.check_value("() * 7")
         self.check_value("() * ()")
